Se agrega rango

+++ src/main/java/edu/fiuba/algo3/modelo/Ciudad.java
@@ -30,7 +30,7 @@ public class Ciudad {
         edificios.add(edificio);
     }
 
-    public double viajar(int velocidad) {
+    public double viajar(double velocidad) {
         return distancia / velocidad;
     }
 }
